Address more comments

+++ build.py
@@ -25,9 +25,6 @@ import sysconfig
 import tarfile
 import tempfile
 
-from ycmd.completers.cpp.clangd_completer import LLVM_RELEASE
-from ycmd.completers.cpp.clangd_completer import GetClangdCommand
-
 PY_MAJOR, PY_MINOR, PY_PATCH = sys.version_info[ 0 : 3 ]
 if not ( ( PY_MAJOR == 2 and PY_MINOR == 7 and PY_PATCH >= 1 ) or
          ( PY_MAJOR == 3 and PY_MINOR >= 4 ) or
@@ -132,11 +129,11 @@ def OnFreeBSD():
 
 
 def OnAArch64():
-  return platform.machine().lower().startswith('aarch64')
+  return platform.machine().lower().startswith( 'aarch64' )
 
 
 def OnArm():
-  return platform.machine().lower().startswith('arm')
+  return platform.machine().lower().startswith( 'arm' )
 
 
 def OnX86_64():
@@ -741,29 +738,21 @@ def EnableJavaCompleter( switches ):
       jdtls_package_name = package_name )
   file_name = p.join( CACHE, package_name )
 
-  if p.exists( REPOSITORY ):
-    shutil.rmtree( REPOSITORY )
-
-  os.makedirs( REPOSITORY )
+  MakeCleanDirectory( REPOSITORY )
 
   if not p.exists( CACHE ):
     os.makedirs( CACHE )
-  elif p.exists( file_name ):
-    with open( file_name, 'rb' ) as existing_file:
-      existing_sha256 = hashlib.sha256( existing_file.read() ).hexdigest()
-    if existing_sha256 != JDTLS_SHA256:
-      Print( 'Cached tar file does not match checksum. Removing...' )
-      os.remove( file_name )
+  elif p.exists( file_name ) and not CheckFileIntegrity( file_name,
+                                                         JDTLS_SHA256 ):
+    Print( 'Cached tar file does not match checksum. Removing...' )
+    os.remove( file_name )
 
 
   if p.exists( file_name ):
     Print( 'Using cached jdt.ls: {0}'.format( file_name ) )
   else:
     Print( "Downloading jdt.ls from {0}...".format( url ) )
-    request = requests.get( url, stream = True )
-    with open( file_name, 'wb' ) as package_file:
-      package_file.write( request.content )
-    request.close()
+    DownloadFileTo( url, file_name )
 
   Print( "Extracting jdt.ls to {0}...".format( REPOSITORY ) )
   with tarfile.open( file_name ) as package_tar:
@@ -786,40 +775,43 @@ def EnableTypeScriptCompleter( args ):
 
 
 def DownloadClangd():
+  LLVM_RELEASE = '7.0.0'
   CLANGD_DIR = p.join( DIR_OF_THIRD_PARTY, 'clangd', )
   CLANGD_CACHE_DIR = p.join( CLANGD_DIR, 'cache' )
   CLANGD_OUTPUT_DIR = p.join( CLANGD_DIR, 'output' )
-  is_64_bit = sys.maxsize > 2**32
   if OnWindows():
-    if is_64_bit:
-      target_name = 'clangd-{LLVM_RELEASE}-win64'
-      check_sum = '2486670cb84c3ea9e9ab3409ebd940ed2c9ddee75adab4745d89df19d029fa54'
-    else:
-      target_name = 'clangd-{LLVM_RELEASE}-win32'
-      check_sum = '533359bb236df7de9f04d6efe3de74a3a6d56f2e1a9154733c076c252d657806'
+    target = [
+      ( 'clangd-{LLVM_RELEASE}-win64',
+        '2486670cb84c3ea9e9ab3409ebd940ed2c9ddee75adab4745d89df19d029fa54' ),
+      ( 'clangd-{LLVM_RELEASE}-win32',
+        '533359bb236df7de9f04d6efe3de74a3a6d56f2e1a9154733c076c252d657806' ) ]
   elif OnMac():
-    target_name = 'clangd-{LLVM_RELEASE}-x86_64-apple-darwin'
-    check_sum = '22431c42404a85c5d0a91b2e3683db08ab8f434c652ec9da2ceb4f168f711579'
+    target = [
+      ( 'clangd-{LLVM_RELEASE}-x86_64-apple-darwin',
+        '22431c42404a85c5d0a91b2e3683db08ab8f434c652ec9da2ceb4f168f711579' ) ]
   elif OnFreeBSD():
-    if is_64_bit:
-      target_name = 'clangd-{LLVM_RELEASE}-amd64-unknown-freebsd11'
-      check_sum = '33d4d399605fce59c80f41bc53cf5fb0f552342c1698d9e6001c9b4edfbbd68c'
-    else:
-      target_name = 'clangd-{LLVM_RELEASE}-i386-unknown-freebsd11'
-      check_sum = '5593e66c6d3e374b0e41b426685d5fd1b50d44ad6301f0148612dc30aa12924a'
+    target = [
+      ( 'clangd-{LLVM_RELEASE}-amd64-unknown-freebsd11',
+        '33d4d399605fce59c80f41bc53cf5fb0f552342c1698d9e6001c9b4edfbbd68c' ),
+      ( 'clangd-{LLVM_RELEASE}-i386-unknown-freebsd11',
+        '5593e66c6d3e374b0e41b426685d5fd1b50d44ad6301f0148612dc30aa12924a' ) ]
   elif OnAArch64():
-    target_name = 'clangd-{LLVM_RELEASE}-aarch64-linux-gnu'
-    check_sum = 'a5a12adc8685754aa1a717fd57d5031805f66d853e1a6fabd5c7f235565cb33a'
+    target = [
+      ( 'clangd-{LLVM_RELEASE}-aarch64-linux-gnu',
+        'a5a12adc8685754aa1a717fd57d5031805f66d853e1a6fabd5c7f235565cb33a' ) ]
   elif OnArm():
-    target_name = 'clangd-{LLVM_RELEASE}-armv7a-linux-gnueabihf'
-    check_sum = 'c776f19f0fd60e5cda550fc862a1c17cc8b1d917d6732931044e72694a824dab'
+    target = [
+      ( 'clangd-{LLVM_RELEASE}-armv7a-linux-gnueabihf',
+        'c776f19f0fd60e5cda550fc862a1c17cc8b1d917d6732931044e72694a824dab' ) ]
   elif OnX86_64():
-    target_name = 'clangd-{LLVM_RELEASE}-x86_64-unknown-linux-gnu'
-    check_sum = '5db646bf789eb5b331ce41275d0575f6f4683921469e33bfaf35f84af72947a8'
+    target = [
+      ( 'clangd-{LLVM_RELEASE}-x86_64-unknown-linux-gnu',
+        '5db646bf789eb5b331ce41275d0575f6f4683921469e33bfaf35f84af72947a8' ) ]
   else:
     print( 'No binaries for your system, please compile it from source.' )
     return False
 
+  target_name, check_sum = target[ sys.maxsize <= 2**32 ]
   target_name = target_name.format( LLVM_RELEASE = LLVM_RELEASE )
   file_name = '{TARGET_NAME}.tar.bz2'.format( TARGET_NAME = target_name )
   download_url = 'https://dl.bintray.com/micbou/clangd/{FILE_NAME}'.format(
@@ -855,12 +847,8 @@ def EnableClangdCompleter( Args ):
     sys.stdout.write( 'Checking for clangd binary...' )
     sys.stdout.flush()
 
-  CLANGD_COMMAND = GetClangdCommand( {} )
-  if not CLANGD_COMMAND:
-    if not Args.quiet:
-      print( "Clangd not found on the path, trying to download it." )
-    if not DownloadClangd():
-      raise Exception( "FAIL: Couldn't download clangd." )
+  if not DownloadClangd():
+    raise Exception( "FAIL: Couldn't download clangd." )
 
   if Args.quiet:
     print( 'OK' )
